// frontend/src/types/relatorioTijolo.ts
// src/types/relatorioTijolo.ts
export interface RelatorioTijolo {
  brancosFurosPares: number;
  brancosFurosImpares: number;
  pretosFurosPares: number;
  pretosFurosImpares: number;
  totalBrancos: number;
  totalPretos: number;
  totalEmInspecao: number;
  totalAprovados: number;
  totalReprovados: number;
  totalDefeituosos: number;
}